bugfix: create the health checker in `access` phase, allow to call `yield` function,   the `healchecheck` need this when creating new object.
